Add support for synchronized lyrics and sidecar (lrc) files
This commit introduces support for reading lyrics from sidecar LRC
files.
Lyrics from LRC files are given higher precedence than embedded lyrics
as they can be easily modified or removed by the user if they wish to
do so. In other words, if a file has both LRC and embedded lyrics, the
LRC file will be given higher precedence.

This commit also adds support for LRC Synchronized Lyrics.
This is done by parsing timed text lyrics from lrc files into a json
based format and storing it in the existing lyrics property of the
track.
The frontend reads this lyrics property and uses a simple heuristic to
detect whether it is json or just plain text and renders the lyrics
accordingly.

Currently, synchronized lyrics are only supported for sidecar LRC
files as that is the de-facto standard for timed lyrics. The other
popular standard is the MP3 ID3 SYLT frame which is specific to
mp3 files. This can be taken up in the future.